Restructure model and update JSON library
