+++ lib/deepl_dart.dart
@@ -3,15 +3,15 @@ library deepl_dart;
 export 'src/deepl.dart';
 export 'src/translator.dart';
 
-export 'src/model/document_handle.dart';
-export 'src/model/document_status.dart';
-export 'src/model/document_translate_options.dart';
-export 'src/model/document_translation_status.dart';
+export 'src/model/translate/document_handle.dart';
+export 'src/model/translate/document_status.dart';
+export 'src/model/translate/document_translate_options.dart';
+export 'src/model/translate/document_translation_status.dart';
 export 'src/model/errors.dart';
-export 'src/model/glossary_entries.dart';
-export 'src/model/glossary_info.dart';
-export 'src/model/glossary_language_pair.dart';
-export 'src/model/language.dart';
-export 'src/model/text_result.dart';
-export 'src/model/translate_text_options.dart';
+export 'src/model/glossaries/glossary_entries.dart';
+export 'src/model/glossaries/glossary_info.dart';
+export 'src/model/glossaries/glossary_language_pair.dart';
+export 'src/model/languages/language.dart';
+export 'src/model/translate/text_result.dart';
+export 'src/model/translate/translate_text_options.dart';
 export 'src/model/usage.dart';
