+++ lib/src/model/glossaries/glossary_entries.dart
