+++ lib/src/model/glossaries/glossary_info.dart
