+++ lib/src/model/glossaries/glossary_language_pair_list_api_reponse.dart
@@ -1,4 +1,4 @@
-import 'package:deepl_dart/src/model/glossary_language_pair.dart';
+import 'package:deepl_dart/src/model/glossaries/glossary_language_pair.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 part 'glossary_language_pair_list_api_reponse.g.dart';
