+++ lib/src/model/translate/document_translation_status.dart
@@ -1,5 +1,5 @@
-import 'package:deepl_dart/src/model/document_handle.dart';
-import 'package:deepl_dart/src/model/document_status.dart';
+import 'package:deepl_dart/src/model/translate/document_handle.dart';
+import 'package:deepl_dart/src/model/translate/document_status.dart';
 
 class DocumentTranslationStatus {
   final DocumentHandle handle;
