+++ lib/src/model/translate/text_result.dart
