+++ lib/src/model/translate/text_result_response.dart
