+++ lib/src/model/translate/translate_text_options.dart
