+++ pubspec.yaml
@@ -11,11 +11,11 @@ environment:
 dependencies:
   collection: ^1.18.0
   http: ^1.1.0
-  json_annotation: ^4.8.1
+  json_annotation: ^4.9.0
   universal_io: ^2.2.2
 
 dev_dependencies:
   build_runner: ^2.4.6
-  json_serializable: ^6.7.1
+  json_serializable: ^6.9.4
   lints: ^4.0.0
   test: ^1.24.4
